+++ WoWDatabaseEditor.Common/WDE.Common/Providers/IItemFromListProvider.cs
@@ -7,6 +7,6 @@ namespace WDE.Common.Providers
     [UniqueProvider]
     public interface IItemFromListProvider
     {
-        System.Threading.Tasks.Task<int?> GetItemFromList(Dictionary<int, SelectOption> items, bool flag, int? current = null);
+        System.Threading.Tasks.Task<long?> GetItemFromList(Dictionary<long, SelectOption> items, bool flag, long? current = null);
     }
 }
